# state field of Task's state machine was changed from AtomicUsize to usize