--- a/glommio/src/task/header.rs
+++ b/glommio/src/task/header.rs
@@ -4,9 +4,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2020 Datadog, Inc.
 //
 use core::alloc::Layout;
-use core::cell::UnsafeCell;
 use core::fmt;
-use core::sync::atomic::{AtomicUsize, Ordering};
 use core::task::Waker;
 
 use crate::task::raw::TaskVTable;
@@ -20,12 +18,12 @@ pub(crate) struct Header {
     /// Current state of the task.
     ///
     /// Contains flags representing the current state and the reference count.
-    pub(crate) state: AtomicUsize,
+    pub(crate) state: usize,
 
     /// The task that is blocked on the `JoinHandle`.
     ///
     /// This waker needs to be woken up once the task completes or is closed.
-    pub(crate) awaiter: UnsafeCell<Option<Waker>>,
+    pub(crate) awaiter: Option<Waker>,
 
     /// The virtual table.
     ///
@@ -39,53 +37,34 @@ impl Header {
     ///
     /// This method will mark the task as closed, but it won't reschedule the task or drop its
     /// future.
-    pub(crate) fn cancel(&self) {
-        let mut state = self.state.load(Ordering::Acquire);
-
-        loop {
-            // If the task has been completed or closed, it can't be canceled.
-            if state & (COMPLETED | CLOSED) != 0 {
-                break;
-            }
-
-            // Mark the task as closed.
-            match self.state.compare_exchange_weak(
-                state,
-                state | CLOSED,
-                Ordering::AcqRel,
-                Ordering::Acquire,
-            ) {
-                Ok(_) => break,
-                Err(s) => state = s,
-            }
+    pub(crate) fn cancel(&mut self) {
+        // If the task has been completed or closed, it can't be canceled.
+        if self.state & (COMPLETED | CLOSED) != 0 {
+            return;
         }
+
+        // Mark the task as closed.
+        self.state |= CLOSED;
     }
 
     /// Notifies the awaiter blocked on this task.
     ///
     /// If the awaiter is the same as the current waker, it will not be notified.
     #[inline]
-    pub(crate) fn notify(&self, current: Option<&Waker>) {
-        // Mark the awaiter as being notified.
-        let state = self.state.fetch_or(NOTIFYING, Ordering::AcqRel);
+    pub(crate) fn notify(&mut self, current: Option<&Waker>) {
+        // Take the waker out.
+        let waker = self.awaiter.take();
 
-        // If the awaiter was not being notified nor registered...
-        if state & (NOTIFYING | REGISTERING) == 0 {
-            // Take the waker out.
-            let waker = unsafe { (*self.awaiter.get()).take() };
+        // Mark the state as not being notified anymore nor containing an awaiter.
+        self.state &= !AWAITER;
 
-            // Mark the state as not being notified anymore nor containing an awaiter.
-            self.state
-                .fetch_and(!NOTIFYING & !AWAITER, Ordering::Release);
-
-            if let Some(w) = waker {
-                // We need a safeguard against panics because waking can panic.
-                abort_on_panic(|| match current {
-                    None => w.wake(),
-                    Some(c) if !w.will_wake(c) => w.wake(),
-                    Some(_) => {}
-                });
-            }
+        if let Some(w) = waker {
+            // We need a safeguard against panics because waking can panic.
+            abort_on_panic(|| match current {
+                None => w.wake(),
+                Some(c) if !w.will_wake(c) => w.wake(),
+                Some(_) => {}
+            });
         }
     }
 
@@ -93,75 +72,11 @@ impl Header {
     ///
     /// This method is called when `JoinHandle` is polled and the task has not completed.
     #[inline]
-    pub(crate) fn register(&self, waker: &Waker) {
-        // Load the state and synchronize with it.
-        let mut state = self.state.fetch_or(0, Ordering::Acquire);
-
-        loop {
-            // There can't be two concurrent registrations because `JoinHandle` can only be polled
-            // by a unique pinned reference.
-            debug_assert!(state & REGISTERING == 0);
-
-            // If we're in the notifying state at this moment, just wake and return without
-            // registering.
-            if state & NOTIFYING != 0 {
-                abort_on_panic(|| waker.wake_by_ref());
-                return;
-            }
-
-            // Mark the state to let other threads know we're registering a new awaiter.
-            match self.state.compare_exchange_weak(
-                state,
-                state | REGISTERING,
-                Ordering::AcqRel,
-                Ordering::Acquire,
-            ) {
-                Ok(_) => {
-                    state |= REGISTERING;
-                    break;
-                }
-                Err(s) => state = s,
-            }
-        }
-
+    pub(crate) fn register(&mut self, waker: &Waker) {
         // Put the waker into the awaiter field.
-        unsafe {
-            abort_on_panic(|| (*self.awaiter.get()) = Some(waker.clone()));
-        }
+        abort_on_panic(|| self.awaiter = Some(waker.clone()));
 
-        // This variable will contain the newly registered waker if a notification comes in before
-        // we complete registration.
-        let mut waker = None;
-
-        loop {
-            // If there was a notification, take the waker out of the awaiter field.
-            if state & NOTIFYING != 0 {
-                if let Some(w) = unsafe { (*self.awaiter.get()).take() } {
-                    abort_on_panic(|| waker = Some(w));
-                }
-            }
-
-            // The new state is not being notified nor registered, but there might or might not be
-            // an awaiter depending on whether there was a concurrent notification.
-            let new = if waker.is_none() {
-                (state & !NOTIFYING & !REGISTERING) | AWAITER
-            } else {
-                state & !NOTIFYING & !REGISTERING & !AWAITER
-            };
-
-            match self
-                .state
-                .compare_exchange_weak(state, new, Ordering::AcqRel, Ordering::Acquire)
-            {
-                Ok(_) => break,
-                Err(s) => state = s,
-            }
-        }
-
-        // If there was a notification during registration, wake the awaiter now.
-        if let Some(w) = waker {
-            abort_on_panic(|| w.wake());
-        }
+        self.state |= AWAITER;
     }
 
     /// Returns the offset at which the tag of type `T` is stored.
@@ -176,7 +91,7 @@ impl Header {
 
 impl fmt::Debug for Header {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let state = self.state.load(Ordering::SeqCst);
+        let state = self.state;
 
         f.debug_struct("Header")
             .field("scheduled", &(state & SCHEDULED != 0))