--- a/glommio/src/task/state.rs
+++ b/glommio/src/task/state.rs
@@ -53,17 +53,6 @@ pub(crate) const HANDLE: usize = 1 << 4;
 /// check that tells us if we need to wake anyone.
 pub(crate) const AWAITER: usize = 1 << 5;
 
-/// Set if an awaiter is being registered.
-///
-/// This flag is set when `JoinHandle` is polled and we are registering a new awaiter.
-pub(crate) const REGISTERING: usize = 1 << 6;
-
-/// Set if the awaiter is being notified.
-///
-/// This flag is set when notifying the awaiter. If an awaiter is concurrently registered and
-/// notified, whichever side came first will take over the reposibility of resolving the race.
-pub(crate) const NOTIFYING: usize = 1 << 7;
-
 /// A single reference.
 ///
 /// The lower bits in the state contain various flags representing the task state, while the upper
@@ -72,4 +61,4 @@ pub(crate) const NOTIFYING: usize = 1 << 7;
 ///
 /// Note that the reference counter only tracks the `Task` and `Waker`s. The `JoinHandle` is
 /// tracked separately by the `HANDLE` flag.
-pub(crate) const REFERENCE: usize = 1 << 8;
+pub(crate) const REFERENCE: usize = 1 << 6;