### What does this PR do?

state field of Task's state machine was changed from AtomicUsize to usize.
States which were used only to ensure thread safety (REGISTRING and NOTIFYING) were removed as not needed.

### Motivation

Removing of usage of membars during changing of a state of the state machine.

### Checklist

[] I have added unit tests to the code I am submitting
[] My unit tests cover both failure and success scenarios
[] If applicable, I have discussed my architecture
[x] The new code I am adding is formatted using `rustfmt`

### Notes from maintainer:

The effects of this are visible on the executor benchmark. Quite consistently I get a 70s speedup in task execution

Before:
```
cost to run task no task queue: 216ns
cost to run task in task queue: 377ns
```
After

```
cost to run task no task queue: 129ns
cost to run task in task queue: 301ns
```

There are similar (albeit a bit less) speed ups in the contended cases for semaphore and local channel.

Very good job, Andrey!