@glommer 2020-11-28
It kind of sucks that github doesn't include the PR description in the merge commit.
That's good information to have in the tree, I wanted the benchmark results to be there =(

Maybe there's a setting for that?